Remove term fetching code and use FetchJsonWebpackPlugin to load term data
The commit removes the school term fetching code and adds an implementation of FetchJsonWebpackPlugin to the webpack configuration. This plugin retrieves school term data during application build and stores it in localStorage for further use. It provides a more optimal approach to term-data fetching and manipulation. In addition, minor changes were made to package.json and package-lock.json to reflect the new dependencies and versions.